"Applied fix from trunk for revision: 808786 " ------------------------------------------------------------------------ r808786 | jleroux | 2009-08-28 09:21:52 +0200 (ven. 28 ao–t 2009) | 14 lignes
Fix a bug reported by J n Valkovic on user ML about about service CreateLead informations in Webtools (actions is not the same as actionsAndSet)
in Webtools, in the ECA section of services info, the I have added a column for fields set.

In availableservices.ftl, I have also replaced
<td><b>${uiLabelMap.WebtoolsIsService}</b> ${condition.isService?default(uiLabelMap.CommonNA)}</td>
by
<#if condition.isService?has_content>
  <td><b>${uiLabelMap.WebtoolsIsService}</b> ${condition.isService}</td>
  <td colspan="2">&nbsp;</td>
</#if>

I think we don't need to show informations (actually always NA) about attributes not used. But as I was not sure everybody would agree, I only did this one as an example. So What do you think, should we do it for all occurrences ?


------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches@809142 13f79535-47bb-0310-9956-ffa450edef68