"Applied fix from trunk for revision: 813702" ------------------------------------------------------------------------ r813702 | jleroux | 2009-09-11 09:54:45 +0200 (ven. 11 sept. 2009) | 3 lignes
A patch from Pradeep Ruhil "This patch is for Improvent in HR module in Global Hr Setting tabs in Paygrade section." (https://issues.apache.org/jira/browse/OFBIZ-2739) - OFBIZ-2739
This issue is that when you select any payGradeId that is already created and then you check the Salary Steps for that payGradeId, it was showing the Salary Steps of all the payGradeId present. So I resolved that issue and now it will display the Salary Step for the selected payGradeId only.

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches@813704 13f79535-47bb-0310-9956-ffa450edef68