"Applied fix from trunk for revision: 833113" ------------------------------------------------------------------------ r833113 | jleroux | 2009-11-05 19:26:27 +0100 (jeu. 05 nov. 2009) | 3 lignes
Change "Geo Point" title to "Point" 
Change the type of relation for Postal Address to GeoPoint from one to many. Because of the elevation field in GeoPoint. A bit scifi for now, but maybe useful one day...)
Thanks to a discussion with Erwan "Relation title not needed on Facility entity" at https://issues.apache.org/jira/browse/OFBIZ-3153 - OFBIZ-3153
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches@833114 13f79535-47bb-0310-9956-ffa450edef68