"Applied fix from trunk for revision: 890245  " ------------------------------------------------------------------------ r890245 | jleroux | 2009-12-14 10:35:15 +0100 (lun. 14 d‚c. 2009) | 2 lignes
A patch from Si Chen "wsdl generated is not correct for Apache Synapse and Visual Studio" (https://issues.apache.org/jira/browse/OFBIZ-3299) - OFBIZ-3299
The WSDL generated by ModelService is not correct for Visual Studio or Apache Synapse, causing Web services not to work properly. The soap:action tag is missing the URI for soapAction and needs to specify that this style is "rpc" for RPC/literal.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches@890249 13f79535-47bb-0310-9956-ffa450edef68