"Applied fix from trunk for revision: 954583" ------------------------------------------------------------------------ r954583 | jleroux | 2010-06-14 20:40:05 +0200 (lun. 14 juin 2010) | 8 lignes
Fix a bug that seems to have been introduced at r934389 for (https://issues.apache.org/jira/browse/OFBIZ-3703) - OFBIZ-3703

This fix a bug in updatePartyContactMech. When you wanted to update only the extension of a phone number it did not get stored. 
I said "that seems to have been introduced" because it's working in stable demo.

I also remove a useless couple of lines at the beginning of the updatePartyContactMech service.

Finally I must say that I find very weird to have extension in PartyContactMech entity. It would be certainly more logical to have in TelecomNumber. Is it really used for anything else than phone number extension?
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@954737 13f79535-47bb-0310-9956-ffa450edef68