"Applied fix from trunk for revision: 958343" ------------------------------------------------------------------------ r958343 | jleroux | 2010-06-27 09:42:46 +0200 (dim. 27 juin 2010) | 2 lignes
The Sales by Store Report was wrong has it did not use a complex alias for quantityOrdered (OrderItem.quantity - OrderItem.cancelQuantity) but only OrderItem.quantity. Thus canceled quantities were not taken into account.
A little drawback is that if there are 0 values sold this values show. But actually it's not a problem since it informs the user that all quantities ordered for this product has been canceled
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@958347 13f79535-47bb-0310-9956-ffa450edef68