"Applied fix from trunk for revision: 967098" ------------------------------------------------------------------------ r967098 | jleroux | 2010-07-23 15:59:16 +0200 (ven. 23 juil. 2010) | 4 lignes
A patch from Sascha Rodekamp "NPE in Financial Account Services" (https://issues.apache.org/jira/browse/OFBIZ-3873) - OFBIZ-3873

NPE in the financial account services when creating a financial account for a store. The error occurred when the accountValidDays is empty. The thru date can not be set.
This patch covers this issue by checking if accountValidDays is empty or not.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@967131 13f79535-47bb-0310-9956-ffa450edef68