"Applied fix from trunk for revision: 931594" ------------------------------------------------------------------------ r931594 | doogie | 2010-04-07 17:47:27 +0200 (mer. 07 avr. 2010) | 6 lignes
If an exact productStoreShipMethId is requested, then do an exact AND
match, not an OR.  Otherwise, both the original exact entity will be
returned, *and* any others that happen to have the same
carrierPartyId/shipmentMethodTypeId pair.  Then, when getFirst is
called, it will effectively be random.  This bug shows up with showing
cost estimates for a shipping method picker during checkout.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@985858 13f79535-47bb-0310-9956-ffa450edef68