"Applied fix from trunk for revision: 990339" ------------------------------------------------------------------------ r990339 | jleroux | 2010-08-28 12:53:03 +0200 (sam. 28 ao–t 2010) | 4 lignes
A patch from Sascha Rodekamp "Xml Ds Dump Iterator Fix - Webtools" (https://issues.apache.org/jira/browse/OFBIZ-3906) - OFBIZ-3906

Issue with the Xml Export: it seems that there is an issue with the each() loop in Groovy, during the export OFBiz throws a lot of warnings which regards to an wrong used iterator. 
So i changed the each() loop into a normal iterator which solves the issue.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@990340 13f79535-47bb-0310-9956-ffa450edef68