"Applied fix from trunk for revision: 1038228  " ------------------------------------------------------------------------ r1038228 | jleroux | 2010-11-23 18:51:13 +0100 (mar., 23 nov. 2010) | 6 lines
A patch from Sascha Rodekamp "Paging for products in category mgr is not correct" (https://issues.apache.org/jira/browse/OFBIZ-4033) - OFBIZ-4033

As you can see at https://demo-trunk.ofbiz.apache.org/catalog/control/EditCategoryProducts?productCategoryId=100
The number of products is 'one' (but there are more than one). It's related to the distinct operator in the EntityFindOptions.

This patch fixes the issue and shows the paging in the right way.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1038233 13f79535-47bb-0310-9956-ffa450edef68