"Applied fix from trunk for revision: 1042411" ------------------------------------------------------------------------ r1042411 | jleroux | 2010-12-05 20:09:57 +0100 (dim., 05 déc. 2010) | 12 lines
A patch from Anne Jessel "Wrong product displayed as associated product" (https://issues.apache.org/jira/browse/OFBIZ-4023) - OFBIZ-4023

To reproduce:
    * add two associated products to a main product, both with "Complementary or Cross-Sell" association type
    * view the main product in ecommerce

Expected behaviour: the two associated products are listed near the bottom of the page
Actual behaviour: two products are listed at the bottom of the page, but only one of them is an associated product. The other is the main product.
The problem is caused by the value of a freemarker variable changing at a point where other freemarker code assumes the variable does not change. The execution path is complicated.
The patch fixes the problem by saving the value of the key variable into another variable, and restoring it later. However I do wonder if there is a better solution.

JLR: I did not try to find a better solution
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1042414 13f79535-47bb-0310-9956-ffa450edef68