Applied fix from trunk for revision: 1052195 ===
OFBIZ-4071 - menu-item's disable-if-empty attribute was not being handled, fixed by using the existing unused method to check if the item should be disabled and if true then applying the disabled style and not rendering the hyperlink portion of the child link (if present).  Based on a slightly modified patch and report from Deepak Dixit, thanks Deepak.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1052197 13f79535-47bb-0310-9956-ffa450edef68