"Applied fix from trunk for revision: 1057519  " ------------------------------------------------------------------------ r1057519 | jleroux | 2011-01-11 09:28:57 +0100 (mar., 11 janv. 2011) | 13 lines
A patch from Paul Foxworthy "findOrdersToPickMove: EntityListIterator not closed if no Picklist generated" (https://issues.apache.org/jira/browse/OFBIZ-4105) - OFBIZ-4105

Go to ordermgr/control/orderview for an order that doesn't have stock ready for picking.
Click on Prink Pick Sheet
You'll get a Pick Sheet PDF with the message "Order not ready for picking, needs stock move"
Look at logs, you'll see an EntityListIterator was created and never closed, so a warning message was logged when the finalize was executed.

The problem is in the findOrdersToPickMove simple method in
applications/product/script/org/ofbiz/shipment/picklist/PicklistServices.xml .

There's a <use-iterator/> for the OrderHeaderAndItemFacilityLocation entity.
After the entity-condition element, there's an <if> element, and the iterator is only used when the "if" condition is false, i.e. the <else> part is executed.
My fix is simply to move the <entity-condition>, complete with <use-iterator>, within the <else> element so the condition is evaluated and the iterator created only when the iterator will be used.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@1057532 13f79535-47bb-0310-9956-ffa450edef68