"Applied fix from trunk for revision: 1058488" ------------------------------------------------------------------------ r1058488 | jleroux | 2011-01-13 11:27:49 +0100 (jeu., 13 janv. 2011) | 3 lines
A patch from Sascha Rodekamp <<Error when "Add All to Cart" or "Add Checked to Cart" viewing an order in ecommerce>> (https://issues.apache.org/jira/browse/OFBIZ-3881) - OFBIZ-3881

The reason for the bug was that the request was called with an additional orderId get parameter. But the orderId was still in the parameter map so the request handler created with these two orderIds a orderId List. These list couldn't be interpreted from the groovy script(because a string was expected). The result was the error message.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1058491 13f79535-47bb-0310-9956-ffa450edef68