"Applied fix from trunk for revision: 1070193" ------------------------------------------------------------------------ r1070193 | jleroux | 2011-02-13 09:54:27 +0100 (dim., 13 févr. 2011) | 5 lines
A patch from Martin Kreidenweis "savePaymentGatewayResponse throws exceptions due to foreign key constraints" (https://issues.apache.org/jira/browse/OFBIZ-4176) - OFBIZ-4176

If you use a CC payment service provider that actually wants to save not only PaymentGatewayResponses, but also PaymentGatewayResponseMessages, saving these to the database actually might fail because they are created in the wrong order. The response and the messages are scheduled as separate async rollback services in processCreditResult. So it's not guaranteed the Response is there before OFBiz tries to add the ResponseMessages. 

Fix for the issue: we created a new savePaymentGatewayResponseAndMessages service, that takes both the Response and the ResponseMessages and writes them to the database in the right order.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1070196 13f79535-47bb-0310-9956-ffa450edef68