"Applied fix from trunk for revision: 1085697" ------------------------------------------------------------------------ r1085697 | jleroux | 2011-03-26 12:24:38 +0100 (sam., 26 mars 2011) | 6 lines
A patch from Mirko Vogelsmeier "Exception if a description for an item on main.ftl in accounting is null." https://issues.apache.org/jira/browse/OFBIZ-4229 - OFBIZ-4229

In the main.ftl of the accounting-module there are a lot of items listed (like paymentMethodTypes, paymentTypes, ...).
If one of these items does not have a description defined, an Exception is thrown.
As a fallback to the description i did add the Id.
Like if paymentMethodType.description doesnt exists, it gets the paymentMethodType.paymentMethodTypeId.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1085700 13f79535-47bb-0310-9956-ffa450edef68