"Applied fix from trunk for revision: 1090952" ------------------------------------------------------------------------ r1090952 | jleroux | 2011-04-11 09:03:32 +0200 (lun., 11 avr. 2011) | 4 lines
A patch from Sascha Rodekamp "Infinite loop when instantiate the delegator using distributetCacheClear" https://issues.apache.org/jira/browse/OFBIZ-3987

We found an infinite loop when instantiate the delegator. This happens when you use the distributetCacheClear .
I moved the init of the dist cache clear to a separate method and call it in the factory which resolves the loop issue.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1090955 13f79535-47bb-0310-9956-ffa450edef68