"Applied fix from trunk for revision: 1100197" ------------------------------------------------------------------------ r1100197 | jleroux | 2011-05-06 15:07:18 +0200 (ven., 06 mai 2011) | 3 lines
Closes <<Caching is used by org.ofbiz.webapp.control.ProtectViewWorker(line 68) for UserLoginAndProtectedView (since changeset r1075350), but the view has set never-cache="true">> reported by Jonas Hoef at https://issues.apache.org/jira/browse/OFBIZ-4243

I wondered if I should not remove the event from the common-controller since I guess this is rarely used. But, as the functionality is available OOTB, I preferred to keep it as is. So I simply remove the never-cache from  UserLoginAndProtectedView view-entity. Since it will almost always be empty, no memory issues and better performance are expected. If I remember well, at the moment I committed this, I checked: the performance impact was already not noticeable (most of the time only one check in the event).
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1100201 13f79535-47bb-0310-9956-ffa450edef68