"Applied fix from trunk for revision: 1175135" ------------------------------------------------------------------------ r1175135 | jleroux | 2011-09-24 11:24:39 +0200 (sam., 24 sept. 2011) | 6 lines
Fix a bug reported by Kiran Gawde at https://issues.apache.org/jira/browse/OFBIZ-4440 "OnePageCheckout breaks when you are logged in with userid that belongs to PartyGroup"

OnePageCheckout breaks when you are logged in with userid that belongs to PartyGroup. But the quickcheckout doesn't break. This is due to the fact that EditShippingAddress has following line which returns null.

person = delegator.findByPrimaryKey("Person", [partyId : party.partyId]);

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1175138 13f79535-47bb-0310-9956-ffa450edef68