"Applied fix from trunk for revision: 1206507" ------------------------------------------------------------------------ r1206507 | jleroux | 2011-11-26 18:38:56 +0100 (sam., 26 nov. 2011) | 7 lines
A patch from Paul Foxworthy "loadSalesInvoiceFact should be post-run" https://issues.apache.org/jira/browse/OFBIZ-4582

See thread at http://ofbiz.135035.n4.nabble.com/Should-loadSalesInvoiceFact-be-run-on-return-instead-of-commit-td3896289.html

The loadSalesInvoiceFact service in applications/accounting/servicedef/secas_olap.xml is set to run asynchronously on (that is, before) a commit. The result is that the service may attempt to read an order that has not yet been committed.

In the patch, the timing has been changed to global-commit-post-run to ensure the order has been committed.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@1206510 13f79535-47bb-0310-9956-ffa450edef68