A patch from Ankit Jain https://issues.apache.org/jira/browse/OFBIZ-OFBIZ-4613 "Vendor tab in "Catalog > Products" is not working."
Some of the tabs like(Geos, Maintenance Meter, Subscription Resources, Workeffort, Orders) are not working because of the same problem, So I am adding a patch for it, instead of adding it in all screens I have added it in the decorator. Also "Orders" tab is not highlighted when selected also fixed it.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1212107 13f79535-47bb-0310-9956-ffa450edef68