"Applied fix from trunk for revision: 1212147" ------------------------------------------------------------------------ r1212147 | jleroux | 2011-12-08 22:56:02 +0100 (jeu., 08 déc. 2011) | 10 lines
A patch from Martin Kreidenweis https://issues.apache.org/jira/browse/OFBIZ-4451

Warning in log when using @renderSubContentCache freemarker macro

When using the @renderSubContentCache freemarker macro, warnings like this appear in the log:

[      GenericEntity.java:314:WARN ] The field name (or key) [caContentAssocTypeId] is not valid for entity [ContentDataResourceView], printing IllegalArgumentException instead of throwing it because Map interface specification does not allow throwing that exception.

Patch: GenericValue.get() will actually not throw an exception but instead log a warning if the field doesn't exist, so now we first check if the field we are trying to access exists.
Checking one of the fields through the ModelEntity should be enough, as all three of them come from the same entity.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@1212152 13f79535-47bb-0310-9956-ffa450edef68