"Applied fix from trunk for revision: 1220298" ------------------------------------------------------------------------ r1220298 | jleroux | 2011-12-18 00:02:01 +0100 (dim., 18 déc. 2011) | 4 lines
A patch from Sergei Biletnikov "Shipping group Instructions and gift messages can not be changed/set if there are several shipping groups" https://issues.apache.org/jira/browse/OFBIZ-4609

Just add more than 1 shipping group for an order, and you notice that it is not possible to set or edit the shipping instructions and gift messages, because the code ignores multiple shipping groups and the html have the same DOM id elements, so only first shipping group can be edited. Of course, it is just a development issue.

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1220299 13f79535-47bb-0310-9956-ffa450edef68