"Applied fix from trunk for revision: 1231444" ------------------------------------------------------------------------ r1231444 | jleroux | 2012-01-14 08:31:15 +0100 (sam., 14 janv. 2012) | 11 lines
A patch from Ankit Jain https://issues.apache.org/jira/browse/OFBIZ-OFBIZ-4629 "The Status of PicklistItem never changes from PICKITEM_PENDING to PICKITEM_COMPLETED."

Ankit 1st provided another patch,I asked:

jleroux: It seems to me that this is already done in methods PackingSessionLine.issueItemToShipment, and VerifyPickSessionRow.issueItemToShipment, which makes sense to me, thought I did only a 5 mins review, so I could be wrong, could you please check?

Ankit: Yes I found the method VerifyPickSessionRow.issueItemToShipment which is doing the same thing.

I found a bug too and also the reason for it, actually when you verify an order by passing "PicklistBinId" then it updates the status of picklistitems, but if we verify an order by passing a orderId then the status of the picklistitem never changes to Complete.

The reason is that when we pass orderId the "picklistBinId" is not get set in the verifyPickSession so its not available at the time "issueItemToShipment" is called so its not changing the status to complete.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@1231449 13f79535-47bb-0310-9956-ffa450edef68