"Applied fix from trunk for revision: 1309715" ------------------------------------------------------------------------ r1309715 | jleroux | 2012-04-05 11:27:24 +0200 (jeu., 05 avr. 2012) | 5 lines
A patch from Paul Foxworthy https://issues.apache.org/jira/browse/OFBIZ-4784 "storeIncomingEmail assumes variable is not null when it might be"

In storeIncomingEmail in CommunicationEventServices.java, the code to assemble the aboutThisEmail debugging message assumes addressesFrom and addressesTo are not null. But the methods that return values for those variables might return null if they encounter an exception.

Removes also uses of UtilValidate.isEmpty, which should only be used for collections and sequences (see OFBIZ-4427).
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1309718 13f79535-47bb-0310-9956-ffa450edef68