"Applied fix from trunk for revision: 1326503" ------------------------------------------------------------------------ r1326503 | jleroux | 2012-04-16 08:27:45 +0200 (lun., 16 avr. 2012) | 5 lines
Fixes a bug reported by Dimitrios Moustos on dev ML

The relationship between lead owner and lead company is not expired when converting the lead.
The reason is that as the partyRelationship which is to be expired is looked up the party group information is taken from partyGroupPartyId, which does not exist.
The correct variable to use is partyGroupId. 
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1326504 13f79535-47bb-0310-9956-ffa450edef68