"Applied fix from trunk for revision: 1338591  " ------------------------------------------------------------------------ r1338591 | jleroux | 2012-05-15 10:19:58 +0200 (mar., 15 mai 2012) | 4 lines
A patch from Pierre Smits "Rates don't start on the first moment of a day and don't last till the latest moment of a day" https://issues.apache.org/jira/browse/OFBIZ-4676

Currently rates for persons have a fromDate that starts on the moment the rate is created but not on the first moment of the day. This increases the difficulty to use rates in calculations. Rates should start on the first moment of a day (eg 01-02-2012 00:00:00.000). Thus rates must end on the last moment (eb 31-01-2012 23:59:59 999) of the day previous to the day a new rate starts.

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1338592 13f79535-47bb-0310-9956-ffa450edef68