"Applied fix from trunk for revision: 1340405  " ------------------------------------------------------------------------ r1340405 | jleroux | 2012-05-19 12:17:23 +0200 (sam., 19 mai 2012) | 5 lines
A patch from Leon for "webtools doesn't treat "Date" field properly" https://issues.apache.org/jira/browse/OFBIZ-4847

If an entity field type is "Date" (YYYY-MM-DD), in edit or create form of webtools, the component of this field is redered as "timestamp".

Update the ViewGeneric.ftl to change the related "shortDateInput" from false to true resolves this problem. 
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1340407 13f79535-47bb-0310-9956-ffa450edef68