"Applied fix from trunk for revision: 1340826  " ------------------------------------------------------------------------ r1340826 | jleroux | 2012-05-20 22:27:23 +0200 (dim., 20 mai 2012) | 21 lines
A patch from Daniel Riquelme "ignore-if-empty in entity-options throw NPE when env-name is in fact empty" https://issues.apache.org/jira/browse/OFBIZ-4843

When specifying the ignore-if-empty attribute inside an entity-constraint in an entity-options a NPE will be thrown whenever the env-name of this entity-constraint is empty.

The error occurs because a null condition is created for the entity-constraint.
When validateSql get called in

org.ofbiz.entity.condition.EntityJoinOperator.validateSql(EntityJoinOperator.java:178)

there is no null checking so when

condition.checkCondition(modelEntity);

gets called an NPE is thrown.

The patch modifies the code in

framework/widget/src/org/ofbiz/widget/form/ModelFormField.java

to avoid the inclusion of a null condition to the entity-options.

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1340827 13f79535-47bb-0310-9956-ffa450edef68