"Applied fix from trunk for revision: 1364052" ------------------------------------------------------------------------ r1364052 | jleroux | 2012-07-21 11:06:03 +0200 (sam., 21 juil. 2012) | 22 lines
A patch from Sumit Pandit for "Order view page - picking slip is throwing exception." https://issues.apache.org/jira/browse/OFBIZ-2810

At https://localhost:8443/ordermgr/control/orderview?orderId=DEMO10091 , In shipping panel at shipment row when click on "Packing Slip" it is throwing an exception in rendering the PDF.

Vikas: This is related to the demo order data for purchase order which is not set up correctly. For more information please refer to OFBIZ-1782
Sumit: Found an issue with demo data.
How to verify -
    Go to order detail page -
    https://localhost:8443/ordermgr/control/orderview?orderId=DEMO10091

    Then in shipment information section click on "Packing Slip" -
    https://localhost:8443/facility/control/PackingSlip.pdf?shipmentId=9999&externalLoginKey=EL564429820623

    Issue cause -
    Missing created date in Data.

    Resolution -
    Find patch in attachment.

jleroux: I crossed this recently and thought that the button should not show until packed. But it was rather coded to show "ERROR: No packages found for this shipment" in the PDF. I think it would be better to not show the button, but it's ok with me.


------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1364053 13f79535-47bb-0310-9956-ffa450edef68