"Applied fix from trunk for revision: 1366960" (conflicts handled by hand) ------------------------------------------------------------------------ r1366960 | jleroux | 2012-07-29 23:49:52 +0200 (dim., 29 juil. 2012) | 13 lines
This closes "Price rule add not correctly" https://issues.apache.org/jira/browse/OFBIZ-4978

steps:
1.login and goto price rule tab
2.add a rule
3.add a condition for the rule
The issue is, the added condition is not my selection. I add condition "Party is Fanse Fan", then add, after success, in the detail list, the last party of the select box be shown on the condition , not the one which I choosed.

jleroux:
* The problem reported by Fanse is now corrected
* Fixes an issue when updating a condition (check is now done on input types and not only on value as in creation where it's not a problem, the copy and adaptation from creation to update service implementation was the reason of this bug

I'm in the process of adding some more check for condition creation, but will do that later (ne feature not only fixes as here)
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1366963 13f79535-47bb-0310-9956-ffa450edef68