"Applied fix from trunk for revision: 1395953  " ------------------------------------------------------------------------ r1395953 | jleroux | 2012-10-09 12:48:41 +0200 (mar., 09 oct. 2012) | 9 lines
A patch from Leon for "Internal notes and shipping notes are lost when input an order" https://issues.apache.org/jira/browse/OFBIZ-5038

When we try to create an order, we can input internal note and shipping note in "option" form. But those information is lost when we come back to "option" setting form or when we create the order at last. 
The estimated delivery before/after date are not correctly set either in "option" form.

1. submit internal notes and shipping notes to order cart
2. display estimated shipping before and after date
3. avoid adding internal notes and shipping notes multiple times if submit options repeatedly
4. make it possible to empty internal note or shipping note
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1395966 13f79535-47bb-0310-9956-ffa450edef68