"Applied fix from trunk for revision: 1524950" ------------------------------------------------------------------------ r1524950 | jleroux | 2013-09-20 10:52:08 +0200 (ven. 20 sept. 2013) | 16 lignes
For "If you clear the last packed item from from a shipment with two items and then complete the pack, you get the exception below" https://issues.apache.org/jira/browse/OFBIZ-5308


If you clear the last packed item from from a shipment with two items and then complete the pack, you get the exception below:

Exception thrown while creating the "newEntity" ....ShipmentPackageContent ... Key ... is not present in table shipment_package

This is happening because org.ofbiz.shipment.packing.PackingSession.clearLine() around line 600 is decrementing packageSeq from 1 to zero.  This causes createPackages() to not generate any ShipmentPackages because of the for statement on packageSeq and therefore the subsequent exception.

I fixed this by just commenting out the block starting with "if(line.packageSeq ==".  I did not submit a patch because I do not fully understand the logic and do not have time now to learn it now.

Hopefully the author will be better equipped to deal with it properly.

Skip

jleroux: the path's author (https://issues.apache.org/jira/browse/OFBIZ-2163) Karim Rahimpur  did not answer. But by analogy with clearAllLines I decided that packageSeq should not be less than 1!
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1533637 13f79535-47bb-0310-9956-ffa450edef68