"Applied fix from trunk for revision: 1529588" ------------------------------------------------------------------------ r1529588 | jleroux | 2013-10-06 09:55:01 +0200 (dim. 06 oct. 2013) | 4 lignes
A patch from  Jyoti Sharma  for "Add Item to approved order" https://issues.apache.org/jira/browse/OFBIZ-5340

While adding an item to an approved order if we select orderItemTypeId it gives an error message.
The error was occurring because in "orderItemTypeId" field, value being passed was that of "reasonEnumId". I am submitting a patch having the correct value. 
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1533741 13f79535-47bb-0310-9956-ffa450edef68