"Applied fix from trunk for revision: 1532342" ------------------------------------------------------------------------ r1532342 | jleroux | 2013-10-15 16:03:54 +0200 (mar. 15 oct. 2013) | 11 lignes
This fixes 
"Addition of item to shopping list is not working" https://issues.apache.org/jira/browse/OFBIZ-4903
"Problem when saving a product to a shopping list" https://issues.apache.org/jira/browse/OFBIZ-4589
I don't pretend it's glorious code. For instance I gave up on trying to have the groovy changes snippets consistent in both components and to reuse the way it's done in party. And the way I secured the indexes remember me when I was hacking in APL. For my defense, it was really a mess... and my time is limited...
Anyway, at least it fixes all the issues I crossed in shopping list editing in both party and ecommerce components:
* NPE on sessions ending
* wrong and failing (NPE in ftl) pagination
* error on pagination links after adding a product

Tested with both 2 and 20 size view on both sides

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1533755 13f79535-47bb-0310-9956-ffa450edef68