"Applied fix from trunk for revision: 1539156" ------------------------------------------------------------------------ r1539156 | jleroux | 2013-11-05 23:13:05 +0100 (mar. 05 nov. 2013) | 12 lignes
A patch from  Christian Carlow for "Ship estimate does not appear on shipping options page of purchase order finalization screen" https://issues.apache.org/jira/browse/OFBIZ-OFBIZ-5388 - OFBIZ-5388

The ship estimate field of the shipping options page of the purchase order finalization screen is not populated with the itemShipGroupEstimate value.

To reproduce:
1.  Create an purchase order to DemoSupplier for product   	
GZ-1000
2.  Finalize the order and click continue until reaching the shipping options page
3.  Set the Shipping Estimate Amount and click Continue
4.  Click the "Options" link to navigate to the previous page and notice that the shipping estimate field is not set to the value submitted in the previous step

This patch sets the value of the shipping estimate field to the ship group estimate.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1539161 13f79535-47bb-0310-9956-ffa450edef68