"Applied fix from trunk for revision: 1552149" ------------------------------------------------------------------------ r1552149 | jleroux | 2013-12-19 00:00:57 +0100 (jeu. 19 déc. 2013) | 3 lignes
Fixes an issue where externalKeyParam was not escaped in link so was useless (this is only in .ftl files)
Actually the right thing to do should be to work on migrating all freemarker templates in backend to widgets and using target-type="inter-app" for those. I know Adrian has opened a Jira for that, next task in my TODO list. 
Without this effort we will never be able to style the backend uniformly, notably in custom projects...
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1552154 13f79535-47bb-0310-9956-ffa450edef68