A patch from Christian Carlow  for "Mandatory and dependent production run links don't navigate to corresponding production run page when clicked" https://issues.apache.org/jira/browse/OFBIZ-5502
When clicking the links associated with madatory or dependent production runs, the corresponding production run page is not loaded. This means its required to go back to the production run search page and lookup the productionRunId manually. I think the links should load the corresponding production run page when clicked.

This simple patch resolves the problem. The <hyperlink> elements in the form widget needed the "link-type" attribute set to "anchor" otherwise it defaults to creating a link that submits a form that automatically sets the productionRunId to the current production run 
workEffortId which causes this issue.

jleroux: tree conflict, applied by hand (not merged from trunk)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1563688 13f79535-47bb-0310-9956-ffa450edef68