"Applied fix from trunk for revision: 1574019  " ------------------------------------------------------------------------ r1574019 | jleroux | 2014-03-04 11:37:01 +0100 (mar. 04 mars 2014) | 16 lignes
A patch from Leila Mekika for "Problems in single form when use-when and position are used together" https://issues.apache.org/jira/browse/OFBIZ-5552

The service "renderSingleFormString" check the previous field position to see if the current field must be displayed after the previous field or on a new line. 

When the previous field has a use-when condition that is not valid, it remains as the currentFormField and the next field is compared with this "not rendered" field. If the position of this previous field is the same as the current field, the current is displayed on the next line.


To reproduce, add in a single form the three fields below :

<field position="1" name="first"><display description="first"/></field>
<field position="2" name="second" use-when="true"><display description="second"/></field>
<field position="2" name="second" use-when="false"><display description="second"/></field>

The field second is rendered on a new line

jleroux: this second version works correctly
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1574021 13f79535-47bb-0310-9956-ffa450edef68