Applied fix from trunk for revision: 1609155 ===
This is a slightly modified version of the work contributed by Simone Viani in OFBIZ-5619: when a service model was cloned (e.g. when runSync/call-service are invoked passing the require-new-transaction flag) the service was executed with use-transaction set to true, even if the called service is defined to not use it. Also added some checks to make sure that if require-new-transaction is true then also use-transaction must be true.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1609156 13f79535-47bb-0310-9956-ffa450edef68