"Applied fix from trunk for revision: 1614366" ------------------------------------------------------------------------ r1614366 | jleroux | 2014-07-29 16:37:48 +0200 (mar. 29 juil. 2014) | 4 lignes
A patch from Pierre Smits for "AddOtherPartyRelationship allows user to select roles the selected party doesn't have" https://issues.apache.org/jira/browse/OFBIZ-5694

When a relationship must be set for a party to another party, the role selector for the party shows all roles available in the data set.
This, however, must be limitied to the roles the party has.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1614369 13f79535-47bb-0310-9956-ffa450edef68