"Applied fix from trunk for revision: 1616940" ------------------------------------------------------------------------ r1616940 | jleroux | 2014-08-09 15:05:05 +0200 (sam. 09 août 2014) | 2 lignes
The updateApprovedOrderItems defines the itemAttributesMap, itemEstimatedDeliveryDateMap and itemEstimatedShipDateMap IN parameters as optional. But if you use this service out of an event context then you cross NPEs
This is because in the context of an event the missing parameters are initialised as empty. This hid the issue so far because updateApprovedOrderItems is only used once, called by an event.
------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1616943 13f79535-47bb-0310-9956-ffa450edef68