"Applied fix from trunk for revision: 1624767"  (conflicts handled by hand) ------------------------------------------------------------------------ r1624767 | jacopoc | 2014-09-13 19:05:03 +0200 (sam. 13 sept. 2014) | 4 lignes
Fixed the regression that Jacques introduced with the work on portOffset when the immutability of ServiceLocation was broken; now the class is thread-safe again.
However the portOffset code, before and after my fix, is really ugly and should be reverted completely: the "8080" and "1099" ports, that are configurable in OFBiz, are hardcoded in the portOffset code and so this code can only work when OFBiz is executed with the default ports.


------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1624812 13f79535-47bb-0310-9956-ffa450edef68