"Applied fix from trunk for revision: 1632801 " ------------------------------------------------------------------------ r1632801 | ashish | 2014-10-18 16:38:46 +0200 (sam. 18 oct. 2014) | 10 lignes
Applied patch from jira issue - OFBIZ-4668 - CommunicationEventAndRole entity: Field with name custRequestDate not found.
==============================================================================
when displaying this portal page:
https://localhost:8443/myportal/control/main?portalPageId=MYPORTAL_EMPLOYEE0&parentPortalPageId=MYPORTAL_EMPLOYEE

this message is thrown to the console:
Message: Roll back error, could not commit transaction, was rolled back instead because of: Error in Service [executeFind]: Error running Find on the CommunicationEventAndRole entity: Field with name custRequestDate not found in the CommunicationEventAndRole Entity
==============================================================================
Thanks Mridul for the contribution and thanks Erwan for creating the issue. 

------------------------------------------------------------------------
?\026

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1633017 13f79535-47bb-0310-9956-ffa450edef68