"Applied fix from trunk for revision: 1650887  " ------------------------------------------------------------------------ r1650887 | jleroux | 2015-01-11 12:25:00 +0100 (dim. 11 janv. 2015) | 7 lignes
A patch from Pierre Smits for "link of dependent task in Task Dependencies of task links Dependencies of Task" https://issues.apache.org/jira/browse/OFBIZ-5815

In the dependencies overview of a task links are provided for the tasks listed.
These links link to the dependencies overview of the task that has been accessed.
The links should point to the dependent task.

jleroux: while testing I had a doubt about Pierre's assertion, but I agree it makes more sense to link to the dependent tasks than to the dependencies overview
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1650889 13f79535-47bb-0310-9956-ffa450edef68