Applied fix from trunk for revision: 1651593 ===
Since ServiceEcaSetField are cached, if a seca definition contains a set operation with an env-name like:
 <set field-name="partyId" env-name="${userLogin.partyId}"/>
then the expanded value (at the time of creation) is stored in the ServiceEcaSetField and then reused even when the seca is re-executed with a different context.
This commit fix the bug by removing the outer if statement "if (UtilValidate.isEmpty(this.value))".
Thanks to Deepak Dixit for spotting this issue.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1651599 13f79535-47bb-0310-9956-ffa450edef68