"Applied fix from trunk for revision: 1657714" ------------------------------------------------------------------------ r1657714 | adrianc | 2015-02-06 00:25:50 +0100 (ven. 06 févr. 2015) | 3 lignes
Revert rev 1622256 to fix a service definition caching problem - https://issues.apache.org/jira/browse/OFBIZ-6060.

I'm reverting this instead of fixing it because the original commit was a first step in an effort to eliminate ThreadLocal variables in the framework (so we can make better use of multi-threading). But since then I have come to the conclusion that it will not be possible without a framework redesign.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1657749 13f79535-47bb-0310-9956-ffa450edef68