Applied fix from trunk for revision: 1660031 ===
Applied patch contributed by Gareth Carter in OFBIZ-5910: it is a fix for a bug introduced by the fix for OFBIZ-3382.
The issue reported in OFBIZ-3382 and the fix committed for it need to be reviewed and possibly reworked but in the meantime this fix will solve the problem with broken links reported recently.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1660032 13f79535-47bb-0310-9956-ffa450edef68