"Applied fix from trunk for revision: 1664696  " ------------------------------------------------------------------------ r1664696 | jleroux | 2015-03-06 19:53:42 +0100 (ven., 06 mars 2015) | 13 lines
A slightly modified patch from Christian Carlow for "date-find in form widget does not persist value" https://issues.apache.org/jira/browse/OFBIZ-5074

Reported by Taher:
Any search form that calls "performFind" and has a "date-find" tag will not persist but it will give correct search results. 
For example, If I go to Order -> Requests and search using any of the date fields, the result will be correct but the field will be blank upon returning results. This behavior affects only dates.
I suspect the main offender here is a bug in performFind service or one of the services it calls (prepareFind or executeFind). I believe that the "queryStringMap" does not contain the date-find fields in its returned value. I did not look heavily into the code but I think "fld" types are not persisted in-between requests.

Christian:
This issue is still relevant and needed. The <date-find> values are not persisted after submission due to MacroFormRenderer.java not evaluating value and value2 correctly. In face, value2 isn't event passed to the definition in framework/widget/templates/htmlFormMacroLibrary.ftl. This is a starting point for this functionality. I said starting point because I'm not sure if it handles default values and other field settings that may need to be applied after my patch.

jleroux: It indeed needs to be improved, at least to keep the related date operator... I don't close yet, but would be better to open a new Jira issue...
I simply formatted the patch (tabs instead of 4 spaces) and used UtilGenerics.checkMap() instead of (Map<String, Object>) 

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1664697 13f79535-47bb-0310-9956-ffa450edef68