A patch from Taher Alkhateeb for "date-find in form widget does not persist value" https://issues.apache.org/jira/browse/OFBIZ-5074
Any search form that calls "performFind" and has a "date-find" tag will not persist but it will give correct search results. 

For example, If I go to Order -> Requests and search using any of the date fields, the result will be correct but the field will be blank upon returning results. This behavior affects only dates.

I suspect the main offender here is a bug in performFind service or one of the services it calls (prepareFind or executeFind). I believe that the "queryStringMap" does not contain the date-find fields in its returned value. I did not look heavily into the code but I think "fld" types are not persisted in-between requests.

jleroux: this is a R12.04 specific patch

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1666648 13f79535-47bb-0310-9956-ffa450edef68