"Applied fix from trunk for revision: 1668314  " ------------------------------------------------------------------------ r1668314 | jleroux | 2015-03-21 21:23:22 +0100 (sam., 21 mars 2015) | 6 lines
A patch from Gil Portenseigne for "Setting the reservation of persons on a routing task isn't reflected in the overview of routing tasks" https://issues.apache.org/jira/browse/OFBIZ-4897 reported by Pierre Smits 

In this patch i add a control on reservPersons and add it to the entry map of updateWorkEffort services.
Being zealous, i'm adding specific labels, existing one where not good for the field they were checking.


------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1668316 13f79535-47bb-0310-9956-ffa450edef68