Applied fix from trunk for revision: 1675852 =========================================================== Fixed the order in which invoice and payment transactions are created.
Payment transactions were being created prior to invoice transactions, It causes confusion for accountants in real world.

It was a seca execution order which was causing the issue on setInvoiceStatus. Moved the trigger on setInvoiceStatus for checkInvoicePaymentApplications and capturePaymentsByInvoice from secas.xml to secas_invoice.xml, so that we do invoice transactions prior to payment.
=============================================================


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1675869 13f79535-47bb-0310-9956-ffa450edef68