"Applied fix from trunk for revision: 1716319  " ------------------------------------------------------------------------ r1716319 | jleroux | 2015-11-25 09:15:42 +0100 (mer. 25 nov. 2015) | 17 lignes
A patch from Anne Jessel for "Entity ECAs not triggered correctly when using Delegator.storeAll() method" https://issues.apache.org/jira/browse/OFBIZ-3847

As reported by  Martin Kreidenweis 
The conditions don't work when updating (not creating) entities using the Delegator.storeAll() method. E.g. the following condition does not work:
<eca entity="Product" operation="create-store" event="return">
        <condition field-name="autoCreateKeywords" operator="not-equals" value="N"/>
        <action service="indexProductKeywords" mode="sync" value-attr="productInstance"/>
</eca>
The indexProductKeywords service is called anyway when the product is updated and the autoCreateKeywords was "N" and stays "N". It works correctly for newly created products. 
The problem is in the method GenericDelegator.storeAll(), where unchanged field values are not passed down to the store() method. The store method calls the ECA engine, which does not receive the unchanged values at all and thus cannot evaluate the EECA conditions correctly. 

Scott's comment:We could consider having EntityEcaCondition.eval() perform a db lookup if any of the fields it needs to evaluate are missing. If they are missing it could fully populate the entity with the missing fields so that a lookup is only required at most once per delegator operation. If no ECA conditions require any missing fields then the lookup wouldn't be performed, so the performance penalty would only be incurred when necessary.


Anne: I've created a patch modeled after Scott's suggestion. I'd appreciate someone reviewing it. It does work for me, however I am concerned that it might fail when the value of a field that is part of the EECA condition is being changed from non-null to null using GenericDelegator.storeAll, because of the way storeAll works. There may be a simple solution I've missed.

jleroux: After reviewing  and waiting for Anne's feedback, I commit because even if in some cases it would not work it's at least better than current situation
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1716322 13f79535-47bb-0310-9956-ffa450edef68