"Applied fix from trunk for revision: 1735385  " ------------------------------------------------------------------------ r1735385 | jleroux | 2016-03-17 11:09:53 +0100 (jeu. 17 mars 2016) | 10 lignes
Fixes "Credit card details cannot be updated" - https://issues.apache.org/jira/browse/OFBIZ-6920

It seems after OFBIZ-3956, credit card details can no longer be updated, for example, if one wishes to update the expiry date of a credit card 
results in following "The Following Errors Occurred: Card Number is erroneous".
This issue can be replicated here - https://demo-trunk-ofbiz.apache.org/partymgr/control/editcreditcard?paymentMethodId=9015&partyId=DemoCustomer

jleroux: The problem was because we are calling createCreditCard map proc from updateCreditCard service after calling updateCreditCard map proc. And when we update the data we don't see the real card number but something like "************1111"

I simply copied the needed check from createCreditCard map proc into updateCreditCard map proc but not the isAnyCard call of course.  Then it will work, but if the user changes also the card number. If she does it will need to be right, because it will not be checked! I can't see any other ways because we can't use a hidden field with the real card number in, it would be insecure. We could introduce a specific encryption/decryption but it seems a bit too much. Notably because the only reason of changing a card number while this card is still valid would be  because the number have been wrongly entered. Then better to delete and create a new one, those data are a bit sensible...

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1735389 13f79535-47bb-0310-9956-ffa450edef68