"Applied fix from trunk for revision: 1744117" ------------------------------------------------------------------------ r1744117 | jleroux | 2016-05-16 21:38:14 +0200 (lun. 16 mai 2016) | 4 lignes
Fixes "Pagination Problem in Find Invoices By Due Date"  - https://issues.apache.org/jira/browse/OFBIZ-7070

Instead of calling the getInvoicePaymentInfoListByDueDateOffset service through the controller, uses redirect-parameter and a groovy script to call the service. This prevents the parameters in the pagination URLs.
The groovy script is needed to prevent the initial search which would always be done by a simple "<service..." call in the screen (daysOffset must not be null for the search to fire).
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1744121 13f79535-47bb-0310-9956-ffa450edef68