"Applied fix from trunk for revision: 1744396" ------------------------------------------------------------------------ r1744396 | jleroux | 2016-05-18 13:52:44 +0200 (mer. 18 mai 2016) | 5 lignes
An enhanced patch from Ingo Wolfmayr for "CategoryContent & ProductContent cachekey problem on multi tenant environment" https://issues.apache.org/jira/browse/OFBIZ-7071

When creating a blog-entry with empty article text, it is not possible to add update the article. Content entry is created but not assigned to contentId of blog-entry. The attached patch fixes this.

jleroux: I extended to all possible cases in "content contex"t and checked that there was no cases outside "content context"
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release15.12@1744400 13f79535-47bb-0310-9956-ffa450edef68